Adaugarea Etapei de umplere este finalizata

+++ src/main/java/com/nicolrom/services/MaterialNoticeService.java
@@ -12,7 +12,7 @@ public interface MaterialNoticeService {
 
     void saveMaterialNotices (Set<MaterialNotice> materialNotices);
 
-    Set<MaterialNotice> getMaterialNoticeSet(Phase phase, List<Integer> materialIds, List<Integer> materialsQuantities);
+    Set<MaterialNotice> getMaterialNoticeSet(Phase phase, List<Integer> materialIds, List<Double> materialsQuantities);
 
     Set<MaterialNotice> calculateMaterialsForPhase(Hole hole, Phase phase);
